pyproject.toml
[tool.poetry]
name = "food-helper"
version = "0.1.0"
description = ""
authors = ["Mateusz Szymaniak <mateuszszymaniak@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.7.2"
asttokens = "2.4.1"
authlib = "1.3.0"
bandit = "1.7.5"
beautifulsoup4 = "4.12.2"
black = "23.7.0"
build = "1.1.1"
cachecontrol = "0.14.0"
certifi = "2023.7.22"
cffi = "1.16.0"
cfgv = "3.4.0"
charset-normalizer = "3.2.0"
cleo = "2.1.0"
click = "8.1.7"
colorama = "0.4.6"
coverage = "7.3.2"
crashtest = "0.4.1"
crispy-bootstrap4 = "2022.1"
cryptography = "42.0.5"
distlib = "0.3.7"
django = "4.2.5"
django-bootstrap4 = "23.2"
django-crispy-forms = "2.0"
django-environ = "0.11.2"
django-extensions = "3.2.3"
django-rest-framework = "0.1.0"
djangorestframework = "3.14.0"
dparse = "0.6.4b0"
dulwich = "0.21.7"
executing = "2.0.1"
factory-boy = "3.3.0"
faker = "19.8.0"
fastjsonschema = "2.19.1"
filelock = "3.12.3"
gitdb = "4.0.10"
gitpython = "3.1.35"
icecream = "2.1.3"
identify = "2.5.27"
idna = "3.4"
importlib-metadata = "7.0.2"
installer = "0.7.0"
isort = "5.12.0"
jaraco-classes = "3.3.1"
jinja2 = "3.1.3"
keyring = "24.3.1"
markdown-it-py = "3.0.0"
markupsafe = "2.1.3"
marshmallow = "3.21.1"
mdurl = "0.1.2"
more-itertools = "10.2.0"
msgpack = "1.0.8"
mypy-extensions = "1.0.0"
nodeenv = "1.8.0"
packaging = "23.2"
pathspec = "0.11.2"
pbr = "5.11.1"
pexpect = "4.9.0"
pkginfo = "1.10.0"
platformdirs = "3.10.0"
poetry = "1.8.2"
poetry-core = "1.9.0"
poetry-plugin-export = "1.6.0"
pre-commit = "3.4.0"
prompt-toolkit = "3.0.39"
psycopg2 = "2.9.7"
ptyprocess = "0.7.0"
pycparser = "2.21"
pydantic = "1.10.14"
pygments = "2.16.1"
pyparsing = "3.1.1"
pyperclip = "1.8.2"
pyproject-hooks = "1.0.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
pytz = "2024.1"
pywin32-ctypes = "0.2.2"
pyyaml = "6.0.1"
rapidfuzz = "3.6.2"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
rich = "13.5.2"
ruamel-yaml = "0.18.6"
ruamel-yaml-clib = "0.2.8"
ruff = "0.3.2"
safety = "3.0.1"
safety-schemas = "0.0.2"
shellingham = "1.5.4"
six = "1.16.0"
smmap = "5.0.0"
soupsieve = "2.5"
sqlparse = "0.4.4"
stevedore = "5.1.0"
tomlkit = "0.12.4"
trove-classifiers = "2024.3.3"
typer = "0.9.0"
typing-extensions = "4.10.0"
tzdata = "2023.3"
urllib3 = "2.0.4"
virtualenv = "20.24.4"
wcwidth = "0.2.6"
werkzeug = "3.0.0"
zipp = "3.17.0"
coreapi = "2.3.3"
coreschema = "0.0.4"
django-filter = "24.2"
djangorestframework-simplejwt = "5.3.1"
itypes = "1.2.0"
openapi-codec = "1.3.2"
pyjwt = "2.8.0"
simplejson = "3.19.2"
uritemplate = "4.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
